Add files via upload
"Background Changer" project with React, Vite, and Tailwind CSS is a dynamic web application that allows users to personalize webpage backgrounds seamlessly. Leveraging the power of React for building interactive user interfaces, Vite for fast and efficient development, and Tailwind CSS for quick and responsive styling, this project combines cutting-edge technologies for a modern web experience.